MxShopProject/requirements.txt
djangorestframework==3.13.1
mysqlclient==2.1.0
Pillow==9.0.1
django-filter==21.1
django-cors-headers==3.11.0
django-simpleui==2022.3.1
djangorestframework-simplejwt==5.1.0
drf-extensions==0.7.1
django-redis==5.2.0
